pyproject.toml
[project]
name = "judge-deployment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
]
